connect dna to ui. fix ui bugs. improve ui implementation

+++ ui/src/graphql/mutation/blockContactMutation.ts
@@ -1,7 +1,7 @@
 import gql from "graphql-tag";
 
 export default gql`
-  mutation BlockContact($username: String, $timestamp: Int) {
+  mutation BlockContact($username: String, $timestamp: Float) {
     blockContact(username: $username, timestamp: $timestamp) {
       agent_id
       timestamp
