+++ ui/src/graphql/mutation/removeContactMutation.ts
@@ -1,7 +1,7 @@
 import gql from "graphql-tag";
 
 export default gql`
-  mutation RemoveContact($username: String, $timestamp: Int) {
+  mutation RemoveContact($username: String, $timestamp: Float) {
     removeContact(username: $username, timestamp: $timestamp) {
       agent_id
       timestamp
