+++ ui/src/graphql/mutation/unblockContactMutation.ts
@@ -1,7 +1,7 @@
 import gql from "graphql-tag";
 
 export default gql`
-  mutation UnblockContact($username: String, $timestamp: Int) {
+  mutation UnblockContact($username: String, $timestamp: Float) {
     unblockContact(username: $username, timestamp: $timestamp) {
       agent_id
       timestamp
